pyproject.toml
[project]
name = "codeground"
version = "0.1.0"
description = ""
authors = [
    {name = "Kojaewoong0504",email = "jaewoong.ko0504@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
pre-commit = "^4.2.0"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
# We use preview style for formatting Black itself. If you
# want stable formatting across releases, you should keep
# this off.
preview = true
